refactor: save/load/update seq moved to useSequencerActions
